refactor(onboarding): 응답 객체 entity가 아닌 dto로 변경
