+++ src/main/java/com/goormgb/be/onboarding/dto/OnboardingPreferenceDto.java
@@ -1,4 +1,4 @@
-package com.goormgb.be.onboarding.dto.request;
+package com.goormgb.be.onboarding.dto;
 
 import com.goormgb.be.onboarding.enums.EnvironmentPref;
 import com.goormgb.be.onboarding.enums.MoodPref;
@@ -9,7 +9,7 @@ import com.goormgb.be.onboarding.enums.SeatPositionPref;
 import com.goormgb.be.onboarding.enums.Section;
 import com.goormgb.be.onboarding.enums.Viewpoint;
 
-public record PreferenceRequest(
+public record OnboardingPreferenceDto(
 	Integer priority,
 	Viewpoint viewpoint,
 	SeatHeight seatHeight,
