+++ src/main/java/com/goormgb/be/onboarding/dto/request/OnboardingPreferenceCreateRequest.java
@@ -2,9 +2,11 @@ package com.goormgb.be.onboarding.dto.request;
 
 import java.util.List;
 
+import com.goormgb.be.onboarding.dto.OnboardingPreferenceDto;
+
 public record OnboardingPreferenceCreateRequest(
 	MarketingConsent marketingConsent,
-	List<PreferenceRequest> preferences
+	List<OnboardingPreferenceDto> preferences
 ) {
 	public record MarketingConsent(Boolean marketingAgreed) {}
 }
