+++ src/main/java/com/goormgb/be/onboarding/dto/request/OnboardingPreferenceUpdateRequest.java
@@ -2,7 +2,9 @@ package com.goormgb.be.onboarding.dto.request;
 
 import java.util.List;
 
+import com.goormgb.be.onboarding.dto.OnboardingPreferenceDto;
+
 public record OnboardingPreferenceUpdateRequest(
-	List<PreferenceRequest> preferences
+	List<OnboardingPreferenceDto> preferences
 ) {
 }
